requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication & Security
# bcrypt 4.1+ broke passlib - pin to 4.0.x
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Background Jobs
redis==5.0.1
rq==1.15.1
rq-scheduler==0.13.1

# Vector DB
qdrant-client==1.7.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# LLM Providers (optional)
openai==1.3.7
anthropic==0.7.8

# Templating
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
